Move GCC/Preprocess to "System" submodule (where they are in Haskell).
