+++ src/syntax/mod.rs
@@ -3,7 +3,5 @@
 
 pub mod ast;
 pub mod constants;
-pub mod gcc;
 pub mod ops;
-pub mod preprocess;
 pub mod utils;
